requirements.txt
beautifulsoup4==4.11.1
fastapi==0.78.0
uvicorn==0.17.6
requests==2.28.0
fastapi-cache2==0.1.8
bonbast==0.5.0